requirements.txt
msal>=1.25.0
msgraph-core>=1.0.0
pydantic-settings>=2.0.0
azure-identity>=1.14.0
azure-monitor-opentelemetry>=1.0.0
opentelemetry-api>=1.20.0

# Microsoft Foundry (Azure AI) - Agent Framework SDK (requires --pre flag)
agent-framework-azure-ai>=0.1.0
azure-ai-projects>=1.0.0b1

# Bot Framework & Teams Integration
botbuilder-core>=4.20.0
botbuilder-integration-aiohttp>=4.20.0
aiohttp>=3.8.0
